refactor: extract ReceiptModal component
